oracle_jobs: Bugfix for auto_drop=TRUE jobs
Oracle allows a definition of Scheduler-Jobs who are
automatically removed after the execution.
They are ignored with this change.

Please update the mk_oracle plugin.

Incompatible Change:
This change has an impact on existing Databases, because Oracle
creates some jobs in SYS with auto_drop. They are ignored with
this Commit.

Change-Id: Id08a953c053941a040ad3fd2e45cd370f3519086