Cleaned up ancient special case of button images
We had two types of internal icons in the past. The ones that start with
icon_* and the ones starting with button_*.  That time the button_*
icons had a highlight variant and a normal image. This was cleaned up
several versions ago. Since then we did not need to distinguish between
the both image types.

This resulted in a lot of duplicate images. These have all been cleaned
up now.

Change-Id: Ice671880eb805fccb9e8573043d7aa96b7a29e3b