6420 agent_3par: Certificate verification can now be enabled/disabled
The special agent is now verifying certificates by default. In case you
have to disable the verification for some reason, you can do this in
the special agent parameters.

However, it is recommended to upload the CA certificate to Check_MK to
enable the special agent to verify the certificate.

Change-Id: I87ef8ff62dc4141500064fc5f3484cbe660c2845