6422 FIX Prevent non-admin users from managing custom host/user attributes
In previous versions it was possible for non-admin users that are allowed
to configure hosts in WATO to also manage custom host attributes.

This has been changed by introducing a new permission "Manage custom attributes"
that is needed in addition to the host / user managing permission. By default
this permission is only given to the "admin" role.

Change-Id: Ife58cae95867904ac3360e4c944a18591fc61b81