Synchronize "omd stop mkeventd" with "omd stop cmc"
Increase the waiting time for "omd stop mkeventd" to 60s (for SIGTERM
after 70s SIGKILL is sent) to leave the mkeventd enough time to finish
gracefully when a high amount of IO is performed on the system. This
is necessary because the mkeventd uses fsync to write the state file
completely before it exits, but fsync does transfer more modified
in-core data than only the state file. This can lead to an increaded
time until the mkeventd stops completely. We observed times close to
16s. To be consistent the times used for SIGTERM and SIGKILL equal to
the times used by the CMC.

Additionally, "omd stop mkeventd" now tests if the pid file is orphanded
before it stops the mkeventd.

CMK-646

Change-Id: I07eeca14c7c2d4bfa5dd371e6fad9087006dbfbf