CMK-615: omd stop on all processes of a site terminates remaining site processes
After a "omd stop" call for all processes of a site stopped the
processes controlled by the init scripts, OMD is now searching for
remaining (uncontrolled) site processes, e.g. still running background
jobs or other processes and sends a SIGTERM to each of them.

It waits up to 5 seconds for them to stop. In case the processes did not
finish in time, it terminates the OMD call with an exit code of 1.

Change-Id: If4da8aa9f8378a74073ffc89813a5f517daaba78