6446 FIX Fixed changing agent type from "No agent" to agent in some cases
With werk #5535 we changed the tag group "Agent type" to be a builtin tag
group. When a pre-1.5 site modified this tag group before updating to 1.5, this
tag group was left untouched.

In this situation hosts that were marked as "No agent" (ping only hosts) were
suddenly set to "Check_MK Agent (Server)" which resulted in error messages
in the monitoring after activating the configuration, because these hosts got
a Check_MK service and tried to contact the agent while there is none on the
affected hosts.

We have now changed the migration logic to extend the locally modified "Agent
type" tag group with the options Check_MK 1.5 needs. This change should make
your "No agent" hosts now stick with the intended configuration.

In case you are affected by this issue,  you may notice that the choices of the
"Agent type" are now extended with some new options which may be confusing in
the first moment. You can now navigate to the "Host tags" page in WATO and edit
the tag group "agent" to sort the new options according to your needs.

Change-Id: I6e5f3031a78022a51c5f0b012e9830b2e6614c12