CMK-767: Add werk 5508
5508 FIX cmk-update-agent.exe: Prevent Agent Updater from running twice

On Windows, when cmk-update-agent.exe is invoked, a lockfile is created to prevent it
from running twice at the same time.

This mechanism did not work properly when cmk-update-agent.exe got invoked
multiple times from different user accounts, because the lockfiles were
created within different directories.

This is now mitigated by creating the lockfile regardless of the current user
within the "temp" directory of the Check_MK Agent installation.

Change-Id: I702e09f7bf673e040a3fb425b417ad87cc361821