6452 SEC Fixed wrong sidebar snapin permission checking
The permission checking of sidebar snapins was not working correctly in all
previous 1.5 versions.

The specific issue was that the default permission of the sidebar snapins were
not set correctly. Especially the master control snapin (which can be used to
globally disable e.g. checking or notifications) was usable even for guest
users by default which is normally only available for administrative users.

As workaround, you could override the default  permissions of the snapins to
make them only available to the intended roles.

Change-Id: I2efac198b69f7f06ee7bfe990b5a832a186286af