oracle_tablespaces: Check for empty filenames
The check goes CRIT when a filename for a file is missing.
This is often the case for Databases after cloning or recreation
of the controlfiles without recreation of the temporary files
for temporary Tablespaces.
The state is always CRIT, because empty temporary Tablespaces
are really bad and a space calculation is not possible.

The plugin from 'Support for Container Database (Multitenant)'
is requirred for the most current SQLs.

Change-Id: I280c71892a8706673897b75996a8e4770deed185