6213 Linux Services
A new check is introduced that checks linux services using 'systemctl --failed'.
The check reports the 'active' attribute of the discovered services and their
description. The state is configurable.

Also a summary of the linux services is given. That check reports the number
of all services and for each service type, i.e. 'active', 'inactive', 'failed'
and other ones.

CMK-117

Change-Id: I2dedec2f37a6287568c50f581cd38f7e97da1308