Moved our pylint configuration to the standard place.
This enables e.g.  the use of Flycheck in Emacs out-of-the-box.  Note that
pylint itself can handle both "pylintrc" and ".pylintrc", but not Flycheck
in its default configuration.  Instead of fighting with tools, just do
things in a standard way.

Fixed some pylint warnings in a modified file on the way.

Change-Id: I4c7c292e6257de7085b722e05294cab8a58b406d