CMK-769: Fix OHM in Win-Agent integration tests
OpenHardwareMonitor was not being built as a prerequisite for
the make target integrationtest for Windows agent. The OHM was falsely
built by configuring it only in Jenkins, which prevented the
integration tests from being run by developers from their own machines.

The correct solution is to build OHM as a prerequisite make target for
integrationtest and not to configure it in Jenkins. This way, the same
setup works for both Jenkins and for developers running the integration
tests from their own machines.

Change-Id: I0b6a3a76ce2c0b21be4ea74e23bba7596e9aeda1