6457 FIX IPMI Management board: Improved error handling when sensors can not be read
When the IPMI sensors can not be read because of an incompatible device, the
error is now handled more gracefully. The other IPMI data, like firmware information,
is fetched even when the sensors can not be read and the exceptions are visible on
the command line when executing e.g. cmk -vd [hostname] or cmk -vvd [hostname].

FEED-931

Change-Id: Id67ab4e70749a6ce8d9f6b71877b0238eaaaf472