6509 Pasting with automatic text splitting in list of input elements
When a dialog is using a self extending list of input elements, like e.g. the
host- and service input fields at the bottom of the "hosts & service parameters"
rule editor, a text can be pasted into the last input field which is then split
by ";" separators. Each of the resulting parts is then added to a dedicated input
field. This is only active for the last (empty) input field.

This makes it a lot easier to populate such dialogs with a longer list of items.

Change-Id: I66c5901f41a1db064e4686bf33d32dcd643efc1d