6453 FIX LDAP: Users with non-ASCII characters in uid could be created incorrectly
When LDAP users with non-ASCII characters in their user ID are not created by the
LDAP sync but during first login, the user data set was stored in a wrong way
(wrong encoded user ID). This could lead to exceptions when the user tried to
login or when an admin tried to work with the user database.

Change-Id: Iea9ba97ec3146435640341e469cc93226fcfce41