6518 FIX Fixed sorting of dual list choices (like network interface types)
When e.g. editing the network interface port types to discovery in the "Network Interface and Switch Port Discovery" ruleset,
the list of port types was initially not sorted correctly. Even when the list was sorted correctly, the sorting was not kept
when moving options between both fields. Both issues have now been fixed.

Change-Id: Id4a44498c156dbe1fde70dcdb3dba43583c68831