6555 Snapins can now be customized
It is now possible to create customized copies of the default sidebar snapins.

To make this possible the snapin developer needs to prepare the snapin for this
feature. The developer needs to specify which parameters can be customized by
the users and make the snapin use the user configuration.

The user can then navigate to "Add snapin page > Custom snapins" to manage his
custom snapins. When editing such a custom snapin the attributes of the snapin
can be edited. The snapins can be shared with other users of the installation.
The sharing logic is equal to the bookmarks, reports and so on.

Once a custom snapin has been created, the user can choose the new dashlet from
the "add snapin page" for adding it to his dashboard.

Change-Id: If0966d9697102f8a1f883bd18908f5ab633a9c6c