6556 Tactical overview and host matrix snapins can be customized
Based on the changes of werk #6555 the Tactical overview snapin can now be
configured to show statistics about custom sets of hosts / services / events.
It can also be configured to show or not show stale objects and failed
notifications.

It is now also possible to create custom copies of the host matrix snapin to
set arbitrary host filters for them which make them able to visualize different
collections of hosts.

Change-Id: I5dd0225339b7d074c7ae7ea7bc67260cd4e94be7