6561 FIX Fixed invalid context button links to "Service globally" and "Graphs globally"
When having a service detail page open, e.g. of a "CPU load" service a user may want to
click on the context button "Service globally" to get a list of all "CPU load" services
in the whole Check_MK setup.

Previous versions were only showing the "CPU load" services of the site the origin service
was located on. The links on the "Service description" worked as intended.

We have now fixed the context button links by removing the "site" filter from the URL.

Change-Id: Id9901ca0b210523c9f402ef862709e2923e55e13