6562 FIX Fixed random "I/O operation on closed file" errors in different situations
When using the LDAP synchronization "I/O operation on closed file" errors could occur in
different places of the GUI with all previous 1.5.0 releases.

These exceptions could happen e.g. during the WATO activate changes or during the user
synchronization. Often it helped to retry the action another time to solve the issue.
Another workaround would be to restart the site apache and retry the action.

Change-Id: I48d092ff596b4a29101e1ca9caa5c9a7cfbdcc64