6709 SEC Fixed possible information disclosure to apache log when editing users
An administrator has the ability to create new users. The credentials of a
newly created user were visible within the HTML of the resulting web page as
GET parameter of various hyperlinks. If one of these links was clicked, the
credentials were stored in the administrator’s browser history and in the access
logs of the server.

CMK-967

Change-Id: I1be7051d97756b1fd4135b032833df18de63eee5