6711 SEC Change Check_MK site umask to prevent "world" access
To prevent Check_MK site files from being read by any local system user the Check_MK
sites now have a umask of 0007 set.

The effect of this change is that new files and directories that are created in
the context of the site user are not accessible by "world" users. These are
local system users that are neither the site user nor members of the site
group.

If you don't like this, you can change the umask back to e.g. 0002 in the file
<tt>~/.profile</tt>.

Change-Id: I9a4a88b3de4937f2fa1fbaa18564930ee08e598a