6709 SEC Fixed possible information disclosure to apache log when ing users
An administrator has the ability to create new users. The entials of a
newly created user were visible within the HTML of the resulting web as
GET parameter of various hyperlinks. If one of these links was clicked, the
credentials were stored in the administrator’s browser history and he
access logs of the server.

CMK-967

Change-Id: I1be7051d97756b1fd4135b032833df18de63eee5