6641 win_license: New check to monitor information about Windows licenses state
This check parses the information provided by the slmgr.vbs command on a
windows host. Information about the License status as well as the time to
expiration are shown.

As a preparation you need to install the agent plugin {win_license.bat}
into your agent's plugins directory.

The check can be configure via the corresponding WATO rule {Windows License}

CMK-1072

Change-Id: I3e3064d78f780c7e62449c1c436d0db52577138d