6716 FIX Add SNMP trap processing log target for deep SNMP trap debugging
A new option "SNMP trap processing" has been added to the Event Console log setting
which tells the underlying SNMP library (PySNMP) to write it's debug output to the
<tt>var/log/mkeventd.log</tt> when set to <tt>Debug</tt>.

Please be careful with this setting. It may result in a lot of log output when you
set it to debug for instances which receive a lot of SNMP traps.

CMK-905

Change-Id: I7d4589434c45edd2b92545b0eb04fdd2eaad00e7