6725 FIX Fixed grouping of service descriptions when view rows are already grouped
When a view is grouped by the host name and it shows only services starting with "Log" of
each host and there is a view grouping for all services starting with "Log" configured,
this resulted in rendering issues. The foldable groups should only fold/unfold services of a
single host (row group) but there were groups spanning over multiple hosts.

Change-Id: I543d93fa193925a50452685fc6c3cefa007cfab7