Refactor help
This poses a great amount of trouble, specially on the web interface.

Depending on context help is translated to help_ or help_txt

under request of many comments in source code it is even refactored to
title as it matches the html attribute title where it is used. Nevertheless
that changes gives quite many problems and maybe needs to be separated in a
different commit.

There is for the moment one API interface function `register_rule` in
watolib.py where the refactoring has not taken place, but is proposed to
skip the check on that function.

Change-Id: Iaf0f65e35a6ee441dcbb10169148a7e8cf5201ed