5511 FIX cmk-update-agent: Improve prelink handling
With werk #5507, we tried to mitigate the impact of prelink on the agent updater binary.
However, it turned out that the corresponding files did not survive every Check_MK Agent
update scenarios.

First of all, please note that we do not recommend the usage of prelink at all, as its benefit is questionable and there are some issues (not limited the the cmk-update-agent binary) that may occur by applying it to a system.

Now, with this werk, the prelink exception for the cmk-update-agent is placed on the target
system in a more robust way.
A blacklist-file at <pre>/etc/prelink.conf.d/cmk-update-agent.conf</pre> is now placed on the target
system regardless of an existing prelink-installation.
If prelink is installed on the target system (i.e. the file /etc/prelink.conf exists),
our blacklist-file will be referenced as an additional prelink configuration file.
If prelink gets installed on the system after the Check_MK Agent, you have to reference
the blacklist-file by yourself. You can do this by adding a line

<pre>
-c /etc/prelink.conf.d/cmk-update-agent.conf
</pre>

to <pre>/etc/prelink.conf</pre>.

However, this mechanism is only available for .rpm and .deb Check_MK Agent pakages.

(Jira Issue: CMK-1056)

Change-Id: Ia65d9dbdc214c405dbe04aa991e43c48c5fa7ab8