5512 FIX cmk-update-agent: Make executable-Format-rule relevant for Linux only
After introducing the binary Format for the agent updater, we added a rule to the
agent updater ruleset that enables the user to choose the format of
the cmk-update-agent executable.

Hence the packaged binary format is only runnable on Linux, this rule now gets ignored
for all other OSes.

Please note: If you want to enable automatic updates on Linux hosts that run on an
architecture other than x86-64 or x86, you have to explicitibly choose to deploy the
python script format, since the binary formats are not runnable on these architectures.

(Jira CMK-1051)

Change-Id: Ia81c1fe383bf32fd1252c225e940dfd5190a17c9