6410 FIX Determine the parent process more reliably
Previously, the parent process was determined by parsing
/proc/{PID}/stat and columns where assumed to be separated by space.
This was unreliable because the filename in the second
column may contain a space as well. If omd was issued e.g. from tmux
the process name "tmux: server" was used. This resulted in a traceback
during e.g. omd rm. Now the psutil module is used to determine the
parent process.

Change-Id: If0a52723a2d95f961973912a4ee3418de90fbef2