6412 FIX mk_docker_container_piggybacked: don't try to execute the check_mk_agent inside a container if no bash is available
If bash is not available in a docker container docker reports the error on stdout.
The mk_docker_container_piggybacked plugin misinterpreted the error as an
available check_mk_agent inside of the container.

Affected hosts using the mk_docker_container_piggybacked plugin have to update
the plugin to the new version to avoid this behaviour.

Change-Id: If851fe33dea6cb861feb9bf416416e0ac802568d