6777 FIX Background job logs were not always shown in job output
In some situations log entries which should be written to the job output,
which is visible from the GUI, was not written to it.

Background: The default logger was based on cmk.log logger which is not
initialized with the GUI log level. The job logger is now based on cmk.gui.log
logger as intended which makes the job log entries now available in the job
progress dialog by default.

Change-Id: I92e65bb7a2ab467fd567c7456c4fbe2761bbfc66