[Check-API] Discovery of single entries and parsed items
It can be used directly in the check info with lambda functions for the
filtering function or as a decorator over the filtering function.

It is possible to setup default parameters

The mysql check plugin is taken as example of use. uptime, cpu_utilization,
lnx_thermal, ups in freq and voltage are used for further examples.

There are tests using the many configurable options

CMK-1013

Change-Id: Iabab6f1b447573a1f0fc5f8900a6bc3c7423ceaa