6778 FIX Additional fix for URL redirect from /[site]/ to /[site]/check_mk/
The werk #6773 changed the behaviour of the URL redirect from <tt>/[site]/</tt> or <tt>/[site]</tt>
to <tt>/[site]/check_mk/</tt>. The change tried to respect custom ports but was breaking some
specific reverse proxy setups.

We have now again changed this setting to completely ignore the protocl, host and port when producing the
redirect answer which makes the client responsible for adding the correct protocol, host and port.

Change-Id: I81df7b425615bc32a3c2464ca82131ff0b011439