6780 FIX Fixed random alert / notify helper crashes on some platforms
The alert and notify helpers of the Microcore could crash "randomly"
after hours of operation.

The crash was visible in form of "ImportError: cannot import name" messages
in the var/log/notify.log or var/log/alert.log.

Due to another problem, which will also be fixed soon, the crash of the
helper process made the Microcore process crash.

As far as we know this happened only on Debian Jessie, CentOS/RedHat 7 and
SLES12 SP1/SP2.

CMK-1085

Change-Id: I50abefc35f89ba688752e1b83a16c6acdc9122ef