6787 SEC Notification spooler: Fixed file path traversal vulnerability
The notification daemon of one site connects to the notification daemon of another site
to exchange notifications between both sites.

The notification daemon was not validating the incoming data correctly which made it possible
for an attacker that has access to the notification sending site to compromise the receiving
site.

Using this vulnerability it was possible to write write files in directories that are writable
by the receiving site user. This could be used to gain access to the site.

CMK-1157

Change-Id: I20cc050a096e3f93827741a9d162c509d575e6fe