Replaced deprecated tuple parameters.
This is a Python-2-only feature, which has been deprecated in PEP 3113 since
2007, see https://www.python.org/dev/peps/pep-3113/.  Some tools (e.g.
Bandit) don't like this feature anymore at all, so it's a good time to
remove it.

This is a totally mechanical change done by 2to3's tuple_params fix,
followed by a few naming improvements here and there.

Change-Id: If42e0921a07ca02b9111d636e3356ee10fe98dd7