6786 SEC Livestatus proxy: Fixed file path traversal vulnerability
The livestatus proxy connects to the livestatus server of remote sites. One task is to
fetch the inventory data of the remote site and replicate it to the master site to make
client accesses faster.

The livestatus proxy was not validating the incoming data correctly which made it possible
for an attacker that has access to the remote sites to compromise the site the livestatus
proxy daemon is running in.

Using this vulnerability it was possible to write write files in directories that are writable
by the liveproxy site user. This could be used to gain access to the liveproxy site.

CMK-1153

Change-Id: Ie4de43fc2f8603ba9f03198384a41291ccee726d