5516 FIX cmk-update-agent: Catch HTTPS redirect
A communication between the Agent Updater and the Check_MK-Server
won't succeed if the communication protocol is configured to HTTP
while the server enforces HTTPS. The reason is that the POST data
sent by the Agent Updater will be lost on the redirect.

So far, this is normal HTTP behavior. Nonetheless, the Agent Updater
now handles this situation by detecting the redirect and directly
sending its data via (unverified) HTTPS.

A warning will therefore be displayed to the user respectively written
to log.

(Jira: FEED-3002)

Change-Id: I0e43e5706cafcb4029c4d0e3bfe1c58fd8a7e70d