FEED-3010: Add tempdir config to agent installation paths rule
The environment variable TEMPDIR can now be set for agent execution
within the ruleset "Installation paths for agent".
The env var will also be set for the start-script of cmk-update-agent
located at /usr/bin. The main reason for this config entry is that
the user can workaround issues with running cmk-update-agent with
the /tmp directory mounted using a noexec-flag.

Change-Id: Iea2f4727b32303e66ed90b382b5b90f43689cda8