Improved view filtering flexibility
Previously we had two different concepts for applying filters to views
and reports. Views could only be filtered by the filters that have
previously been configured by the view owner. In reports all available
filters for a data type could be used.

The view logic has now been changed to fit the report logic. It is now
equal to the report processing logic.

Filters now have this logic:

<ul>
<li>When you create a view without adding a filter the view will not
be filtered when accessing the view. There will be no filter dialog
offered via the GUI. New: The user may now provide a filter related
HTTP variable via URL to filter the view data.</li>
<li>To make a filter available to the user, you need to add this
filter to the view using the view editor.</li>
<li>To pre-filter a view you may add initial values to the filters
that are selected using the view editor.</li>
</ul>

This change was needed to support the configurable sidebar snapin
feature. It needs to be able to use arbitrary filters the user has
configured using the snapin.

Added some tests for the filter selection logic and cleaned up the
related functions a bit to reduce duplicated logic.

CMK-437

Change-Id: I68240571d1e2bb14859481b738fd67f58d8a4577