6673 FIX Checks from mkps did not always replace existing checks
Check_MK precompiles check plugins, so that they can be faster reused later on.
The mechanism which detects if a plugin needs to be precompiled again, missed the fact
that files extracted from mkps may have an older timestamp than the precompiled one.
This also applies to old check files which are moved into the local checks dir.

Change-Id: I86cba177aca54f5273201300a3e1e5b21dcc35af