6648 VictorOPS: Notification plugin for VictorOPS
It is now possible to configure Check_MK notifications to be sent to
VictorOPS by using their REST API integration.

In VictorOPS the REST integration for alerts needs to be enabled and the
routing keys defined. Then in Check_MK for every routing key a notification
can be set up under WATO->Notifications. Copy your VictorOPS REST API URL
with the corresponding routing key in the URL. Optionaly to include the URL
of Check_MK's enable the URL prefix field. (VictorOPS does not support
links in their timeline at the time of this writing, thus the full URL is
displayed in the message)

CMK-1187

Change-Id: If002982cde6fe4ecab6235fcde6468e53b7eb8af