Host attributes can now be locked individually
This is will be needed for the dynamic configuration but is not limited
to this feature. It may be used in other situations where some
attributes of a host / folder are controlled by another tool instead of
manual configuration and manual editing is not permitted.

The locking is done for each object individually. Locked attributes can
not be edited using the GUI anymore.

The lock can currently only be set or removed using the Web API.

Adding a lock is done by setting the "locked" attribute with a data
structure like this:

<pre>
(...)
"locked": {
    "locked_by": ("site_id", "dcd", "connection-id"),
    "attributes": [
        "alias",
        "ipaddress",
        "ipv6address",
    ],
},
(...)
</pre>

The structure above locks the host attributes alias, ipaddress and
ipv6address.

The locked_by attribute is the Check_MK installation global identity of
the component that set the lock. This is either used as information for
that component that this object belongs to it and tells other components
that they are not allowed to edit these attributes.

CMK-1144

Change-Id: I70c8a15d42a0b508a84775104520d206ae84476b